requirements.txt
BeautifulSoup==3.2.1
beautifulsoup4==4.4.1
cffi==1.6.0
click==6.6
colorama==0.3.7
console==0.1.1
cryptography==1.3.2
cycler==0.10.0
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.9.5
django-excel==0.0.4
djangorestframework==3.3.3
enum34==1.1.6
et-xmlfile==1.0.1
gunicorn==19.6.0
httpie==0.9.3
idna==2.1
ipaddress==1.0.16
jdcal==1.2
matplotlib==1.5.1
mysqlclient==1.3.7
ndg-httpsclient==0.4.0
numpy==1.11.0
openpyxl==2.3.5
Pillow==3.3.0
plotly==1.11.0
psycopg2==2.6.2
py==1.4.31
pyasn1==0.1.9
pycparser==2.14
pyexcel==0.2.1
pyexcel-io==0.1.0
pyexcel-webio==0.0.6
Pygments==2.1.3
pyOpenSSL==16.0.0
pyparsing==2.1.4
pytest==2.9.1
python-dateutil==2.5.3
pytz==2016.4
requests==2.6.0
six==1.10.0
static3==0.7.0
texttable==0.8.4
virtualenv==15.0.2
whitenoise==3.2
xlrd==0.9.4
XlsxWriter==0.8.7
xlwt==1.0.0
